paging: widen batch_try_map_user_pages_with_prot to u64 mask
The primitive returned a u32 bitmap, capping per-call batches at 32
pages.  apply_prefault_template for busybox (~260 contiguous entries)
was running 9 batches; widening to u64 drops that to 5 batches —
~45% fewer leaf-PT traversals + DSB/ISB fences.

Updated the apply_prefault_template scratch array from 32 to 64 to
take advantage.  prefault_small_anonymous and init-stack batch
paths are bounded by smaller limits (≤8 anon pages, 1-32 stack
pages) so the 32-cap there wasn't the bottleneck — left as-is.

Three run-to-run variance (fork_exit 23-25 us, exec_true 47-49 us,
shell_noop 66-69 us) puts the per-fix saving below the noise floor,
but the primitive is cleaner and headroom is there for future
callers with bigger batches.

Contracts 159/159, threads-smp 14/14.

+++ platform/arm64/paging.rs
@@ -1089,7 +1089,8 @@ impl PageTable {
 
     /// Batch-map contiguous user pages, traversing the page table hierarchy only
     /// once per leaf PT (2MB region) instead of once per page.
-    /// Returns a u32 bitmask: bit i set = page i was mapped.
+    /// Returns a u64 bitmask: bit i set = page i was mapped.  Caller
+    /// must pass `count <= 64`; larger batches should be chunked.
     #[inline(always)]
     pub fn batch_try_map_user_pages_with_prot(
         &mut self,
@@ -1097,9 +1098,9 @@ impl PageTable {
         paddrs: &[PAddr],
         count: usize,
         prot_flags: i32,
-    ) -> u32 {
+    ) -> u64 {
         let attrs = prot_to_attrs(prot_flags);
-        let mut mapped: u32 = 0;
+        let mut mapped: u64 = 0;
         let mut i = 0;
 
         while i < count {
@@ -1126,7 +1127,7 @@ impl PageTable {
                     core::sync::atomic::Ordering::AcqRel,
                     core::sync::atomic::Ordering::Relaxed).is_ok()
                 {
-                    mapped |= 1 << i;
+                    mapped |= 1u64 << i;
                     let vaddr_val = start_vaddr.value() + i * PAGE_SIZE;
                     crate::flight_recorder::record(
                         crate::flight_recorder::kind::MAP_USER,
